CIV-640 Refactor to change home page from /home to /dashboard
