+++ src/main/routes/features/response/citizenDob/citizenDobController.ts
@@ -1,7 +1,7 @@
 import * as express from 'express';
 import {CitizenDob} from '../../../../common/form/models/citizenDob';
 import {Validator} from 'class-validator';
-import {DOB_URL, ROOT_URL, AGE_ELIGIBILITY_URL} from '../../../../routes/urls';
+import {DOB_URL, DASHBOARD_URL, AGE_ELIGIBILITY_URL} from '../../../../routes/urls';
 import {Respondent} from '../../../../common/models/respondent';
 import {Claim} from '../../../../common/models/claim';
 import {AgeEligibilityVerification} from '../../../../common/utils/ageEligibilityVerification';
@@ -14,7 +14,7 @@ function renderView(res: express.Response, form: CitizenDob): void {
 
 function redirectToNextPage(req: express.Request, res: express.Response, dob: Date) {
   if (AgeEligibilityVerification.isOverEighteen(dob)) {
-    res.redirect(ROOT_URL);
+    res.redirect(DASHBOARD_URL);
   } else {
     res.redirect(AGE_ELIGIBILITY_URL);
   }
