+++ src/main/routes/features/response/citizenPhoneNumber/citizenPhoneController.ts
@@ -1,6 +1,6 @@
 import * as express from 'express';
 import {CitizenTelephoneNumber} from '../../../../common/form/models/citizenTelephoneNumber';
-import {CITIZEN_PHONE_NUMBER_URL, ROOT_URL} from '../../../urls';
+import {CITIZEN_PHONE_NUMBER_URL, DASHBOARD_URL} from '../../../urls';
 import {ValidationError, Validator} from 'class-validator';
 import {Respondent} from '../../../../common/models/respondent';
 import {Claim} from '../../../../common/models/claim';
@@ -32,7 +32,7 @@ router.post(CITIZEN_PHONE_NUMBER_URL,
       claim.legacyCaseReference = 'phone-number';
       const draftStoreClient = req.app.locals.draftStoreClient;
       draftStoreClient.set(claim.legacyCaseReference, JSON.stringify(claim)).then(() => {
-        res.redirect(ROOT_URL);
+        res.redirect(DASHBOARD_URL);
       });
     }
   });
