--- src/main/routes/home.ts
@@ -1,9 +1,0 @@
-import * as express from 'express';
-
-const router = express.Router();
-
-router.get('/', (req, res) => {
-  res.render('home');
-});
-
-export default router;
