+++ src/main/routes/urls.ts
@@ -6,7 +6,7 @@ export const SIGN_OUT_URL: Urls = '/logout';
 export const CASES_URL: Urls = '/cases';
 export const DASHBOARD_URL: Urls = '/dashboard';
 export const CITIZEN_PHONE_NUMBER_URL: Urls = '/citizen-phone';
-export const ROOT_URL: Urls = '/home';
+export const ROOT_URL: Urls = '/';
 export const DOB_URL: Urls = '/your-dob';
 export const AGE_ELIGIBILITY_URL: Urls = '/eligibility/under-18';
 export const UNAUTHORISED_URL: Urls = '/unauthorised';
