+++ src/main/views/webpack/js.njk
@@ -1,1 +1,1 @@
-<script src="/main-dev.js"></script>
+ <script src="main-dev.js"></script>
