+++ src/test/unit/routes/features/dashboard/dashboardController.test.ts
@@ -1,9 +1,11 @@
 import request from 'supertest';
 
-import {app} from '../../../main/app';
+import {app} from '../../../../../main/app';
 import config from 'config';
-jest.mock('../../../main/modules/oidc');
-jest.mock('../../../main/modules/draft-store');
+import {DASHBOARD_URL} from '../../../../../main/routes/urls';
+import {CIVIL_SERVICE_CASES_URL} from '../../../../../main/app/client/civilServiceUrls';
+jest.mock('../../../../../main/modules/oidc');
+jest.mock('../../../../../main/modules/draft-store');
 const nock = require('nock');
 
 const agent = request.agent(app);
@@ -16,7 +18,7 @@ function authenticate() {
     });
 }
 
-describe('Home page', () => {
+describe('Dashboard page', () => {
   const idamUrl: string = config.get('idamUrl');
   const citizenRoleToken: string = config.get('citizenRoleToken');
 
@@ -25,15 +27,18 @@ describe('Home page', () => {
       .post('/o/token')
       .reply(200, {id_token: citizenRoleToken});
     authenticate();
+    nock('http://localhost:4000')
+      .post(CIVIL_SERVICE_CASES_URL)
+      .reply(200, {});
   });
 
   describe('on GET', () => {
-    test('should return sample home page', async () => {
+    test('should return dashboard page', async () => {
       await agent
-        .get('/')
+        .get(DASHBOARD_URL)
         .expect((res) => {
           expect(res.status).toBe(200);
-          expect(res.text).toContain('Default page template');
+          expect(res.text).toContain('Claims made against you');
         });
     });
   });
